src/privategpt/services/ui/requirements.txt
streamlit>=1.46
requests>=2.31
PyPDF2>=3.0
python-docx>=1.1
pymdown-extensions>=10.0 